+++ examples/eeprom/Eeprom2/Eeprom2.cpp
@@ -62,9 +62,9 @@ static uint8_t eeprom_buffer[EEPROM_BUFFER_SIZE];
 
 using namespace eeprom;
 
-static void trace_eeprom(FormattedOutput<OutputBuffer>& out, uint16_t address, uint8_t loops = 1)
+static void trace_eeprom(FormattedOutput<OutputBuffer>& out, uint16_t address, uint16_t loops = 1)
 {
-	for (uint8_t i = 0; i < loops; ++i)
+	for (uint16_t i = 0; i < loops; ++i)
 	{
 		out.width(4);
 		out << address << ": ";
